Update laser.dm
Changes laser cannon fire delay to be 8 instead of 20.  Doubling cannon damage.  This makes it 2/3s more effective than the laser, instead of the laser being >2/3s more effective than the laser cannon.  Prior to bay gun changes it was a full 2x as effective.  I think this is a pretty good range?  -Duck