Update special.dm
Cuts plasma pistol total shots in half (Down to 5) in exchange for letting it shoot as fast as a submachine gun(Firedelay 0).  Lowers weight-class so the pistol (it is a pistol.) is actually small and not the same size as a full laser.  The aim is for a small, very deadly gun that only has enough charge to crit one person -if- you land every shot.  Compare with the standard laser gun, which has enough to crit 3 people and gives you a lot more leeway to miss.  Hoping for an effective, easy-to-use self-defense and/or concealed assassination-y type weapon.  It feels silly when you're arming the crew against the wizard and can only give them fully automatic bulletguns or giant laser cannons.  -Duck.