Update force.dm
Wizard staff is currently a 20-damage laser that can do shrapnels and embeds and does a little knockback.  This bumps it up to be on-par with a laser, plus the shrapnel and knockback.  It is, of course, and always will be one of the weakest choices a wizard can make.  Because you could just teleport into the armoury and steal their laser, or steal the captain's laser instead of wasting a precious spell-slot.  But at least it's not a joke now.