Adds Chef's Special poison stolen from /vg/station
You eat like five units and the poison is invisible.  150-180 ticks later, you drop dead instantly.  You'll probably want to rename it / change the messages and maybe bump up the time it takes for someone to drop dead a bit, since if this gets added as a traitor item, I'm assuming more people than the chef will get it.  It'd be great for the bartender, chef, botanist.  Doctors.  We can dump it in here without commenting it out because there's no way to get it right now.