Preliminary drug-induced deculting.
Uncomment if it's not awful.  Notes from other admins that I think would be really cool?  If a more experienced coder could make it so that cultists are only deculted and uncultable while the drug is in their system, and after it wears off, they become cultists again, that would be amazing.  Thanks.  Code was stolen straight from holy water, so it should work with reagents, but.  There's an if already there and two things under a thing make me nervous.  -Duck.