Adds clonetox, buffs lexorin
Lexorin: 34 seconds to unconsciousness, 6 minutes and 10 seconds to death.
Mutagen: 28 seconds to unconsciousness, 4 minutes and 21 seconds to death.
Destabilizer: 36 seconds to unconsciousness, minute and 16 seconds to death.
Unfortunately, the only thing that's ever really mattered is the knockout time.  Cloneloss(5) drops the knockout time to a minute and 20 seconds because cloneloss doesn't cause much pain, apparently.  Gave minor buff to Lexorin because it was worse than mutagen (Six minutes to kill after the spike is a long time, and it takes plasma.  Might as well make it not awful?)