Kocasslani Release
Todo: Make white only eyes.
Todo #2: Make special blood that is called K, it'll be Kocasslani
special it will be coloured blue and won't be able to transfer to other
types. (It needs to change the color of blood bags to Blue, the sprites
are in the bloodpack.dmi)
Todo #3: After the blue blood release, get a blue overlay_husk and
change the blood color of kocasslani to blue (simply uncomment the flesh
and blood color)