Gravity Generator
- Addition of Gravity Field Generators
- Added animation for when floating
- Toggle Gravity in Secrets menu only works when a Grav Gen is on the
station
- New debug verb to reset the gravity on the whole server

Because of the way this has been implemented, any Grav Gen on any zLevel
will affect the station level only. I would need to rewrite the whole
movement/gravity system as well.
Something I think can be saved for when we switch from station to
outpost.

Mapping info:
At the start of round there doesn't need to be one on the station as the
original gravity system will be running when the round starts.  As soon
as a Grav Gen has been spawned and it has started up (Done automatically
if already on map or spawned using the /station/admin version) It will
take over all gravity simulation.

The main object:
/obj/machinery/gravity_field_generator/main

The one to be place on a map at round start:
/obj/machinery/gravity_field_generator/main/station

The one admins should spawn if they need/want to spawn one:
/obj/machinery/gravity_field_generator/main/station/admin