Fixes #458
Alters the EMP code so that mechas are actually affected by the exosuit
computer commands.
Anything that isn't a Gygax or a Durand will be instantly scuttled, with
the two surviving mechas losing 70-99% of their cell capacity.
Additionally, a small (1 tile) empulse is emitted from the position of
the mecha, disabling headsets and the like.