Adding NanoUI Door Interface
This is a big changelist. This adds the nanoui door interface.
This also adds the concept of the wires datum, which necessitated
modifying the Topic selection routine for machinery, this then required
me to reorganize click code a little.

Almost all of this code is taken from the Paradise server, and I believe
most of it was originally from Baystation.