Adminhelp Dibs!
Adds a dibs feature for claiming adminhelps.
* admin.dm modified with a new get_options_bar state
* admin\topic.dm modified with the new href
* adminhelp.dm modified to include a new adminhelped var state, and to
toss around the information available

When an issue is adminhelped, the links displayed are "?", "DIBS", "PP"
and "JMP".
Following that, an issue can be claimed by clicking DIBS, at which point
both admins and mods are informed of who and what. After DIBS is called,
the tending staff member gets additional options for the target mob and
the player gets a response, saying that their issue will be handled by
so and so. Beyond that, every works as normal.

If DIBS is bypassed with a regular PM to the target, it'll revert
adminhelped to 0, and still give the massive red block of, "Click the
admin's name to reply, please!" so that feature is still intact.