Fixes #586
Adds a "Toggle lid" option to the UI of the DNA scanner. This is mainly
an aesthetic enhancement, making the system completely usable from the
UI, so you don't need to faff with right clicking/verbing.

It also revises the check for the toggle_lid verb: basically, silicons
and anyone else living, who can open the UI, can now toggle it. Makse
sense if it's mechanized and a button.