News Update
Pulls the proper news from the SQL, in accordance with the newer system
centered around the website.

Note that uploadIP, uploadtime and approvetime are specific items for
us. If you don't have a website form for submitting them, then you're
free to not use them (just comment out during execution).

I *was* going to add the newscaster functionality to the PDAs. But. The
code spidered a little too far for me to bother with the present
situation. Sorry v.v