Life Support Enabled Operating Table - Initial Commit Adds an advanced surgery table which enables you to keep a patient alive, even if they're lacking a heart. Also includes changes to the operating computer, making it run off of NanoUI, and makes both the advanced and normal surgery tables constructable with materials from RnD and the medical bay.
The advanced surgery table will require cryogenic freezing fluids to keep in operation, so either cryoxadone or clonexadone will work (though clonexadone will last you longer). You also have the capacity to add an air tank to the machinery and a bloodbag. The operation of both can be automated. The internals will be toggled as soon as you enable life support, and the blood bag will be connected if a patient is below 90% bloodlevel, if the respective systems are made to be automatic. Both can also be manually enabled and disabled.

----

To construct a surgery table, you will require:
* Circuit Board (Operating Table) - Biotech 2, Data theory 1
* Cable Coil
* Capacitor
* Health Analyzer

To construct an advanced surgery table, you will require:
* Circuit Board (Life Support Enabled Operating Table) - Biotech 4, Data theory 3
* Cable Coil
* Capacitor
* Health Analyzer
* Beaker (internal chemical store size dependant on beaker used)
* Advanced Sensor Module

Simple add them to a machine frame, as you would with any other machinery item.

----

A short list of to-dos remains:
* Sprites
* Investigate brain transplants/removal
* Emagged features